Added the ability to fetch repository-logs at runtime. Using runrepositoryfetch.bat, the user is now able to start codeswarm alternatively with a swing-frame.
This frame asks the user for a repository location, a username and a password for the repository.
It is able to connect to any svn-repository, since it uses svnkit to fetch the repository-log.
The repository-log gets fetched and converted to the xml-format, codeswarm expects as input.
The xml-file is cached until the latestRevision of the remote repository changes.
See javadoc for detailed information.